Draft HTTP Response model (#18)
