+++ core/src/main/scala/zio/http/model/Cookie.scala
@@ -18,4 +18,15 @@
 
 package zio.http.model
 
-case class Header(name: String, value: String)
+import java.time.LocalDateTime
+
+// https://tools.ietf.org/html/rfc6265
+final case class Cookie(
+  value: String,
+  domain: Option[String],
+  path: Option[String],
+  expires: Option[LocalDateTime],
+  maxAge: Option[Long],
+  secure: Boolean,
+  httpOnly: Boolean
+)
